Add files via upload
Created individual bash scripts for all categories. Main UnixProject bash script isnt fully bug free. Added all commands with YAY. If you wish to do with pacman you can, however after researched I noticed yay is better, you can always change if you wish. Didn't add if statement for if available space > required space, information about all the choices in each bash file, and an echo for if the install  was successful or not. if the user does not have yay, (or pacman), i encouraged for them to install it, however have a choice for them to select to install for them.